rocker: add dma register macros, and structures
Translate the rocker spec to header file definitions.  This patch includes the
interrupt status register set, as well as the dma/completion/command/event
descriptor ring entry structures.

Note also that this patch updates the spec to clarify the alignment needs to
individual structures.  The padding in the structures was intended to ensure
that each instance ended on an 8 byte boundary.  In an effort to be more
explicit, the documentation has been updated to state that each structure
instance in a dma buffer must be 8 byte aligned, and is implemented in the
header file using an __atttribute__((aligned (8))) directive.

Signed-off-by: Neil Horman <nhorman@tuxdriver.com>
CC: sfeldma@cumulusnetworks.com
CC: Jiri Pirko <jiri@resnulli.us>
Signed-off-by: Scott Feldman <sfeldma@cumulusnetworks.com>