rocker: Rename "FLOW" world to "OF-DPA"
Since the current flow implementation is based on OF-DPA, it makes sense to use
a specific name rather than the generic "flow" name.  That leaves room for
other flow-like worlds in the namespace.  Also, in subsequent patches, some of
the lower-level flow tables and functions will be shared by l2l3 world and
of-dpa world, so prepare for that move now by moving of-dpa-specific parts to
OF-DPA names.

No functional changes here; just renaming.

Signed-off-by: Scott Feldman <sfeldma@cumulusnetworks.com>