rocker: split of_dpa and generic flow-processing code
No functional changes.  Split of-dpa-specifc code into new of_dpa.[hc].  What
remains in flow.[ch] is a generic flow-processing system, oe engine.  Which
flow tables are defined and the flow pipeline is outside the generic code.
OF-DPA, for example, prescribes a flow procssing pipeline using OF flow tables,
but the basic alogrithm for matching a pkt against a flow table, or performing
group actions can be pushed to the generic code and shared by other models,
such as traditional L2/L3 model.

It looks like a lot of change, but really is just a partionaing of existing
code into generic and specific modules.

Signed-off-by: Scott Feldman <sfeldma@cumulusnetworks.com>