rocker: flatten of-dpa flow tlvs
Remove one level of TLV nesting for OF-DPA flow cmds.  Previously, TLVs where
nested as:

	CMD_TYPE
	CMD_INFO
		OF_DPA_TABLE_ID
		OF_DPA_COOKIE
		OF_DPA_HARDTIME
		...
		OF_DPA_INFO
			OF_DPA_INFO_xxx
			OF_DPA_INFO_yyy
			...

Now nesting is:

	CMD_TYPE
	CMD_INFO
		OF_DPA_TABLE_ID
		OF_DPA_COOKIE
		OF_DPA_HARDTIME
		...
		OF_DPA_xxx
		OF_DPA_yyy

The original nesting was an artifact of having per-table nest, but since table
ID is present, the context of OF_DPA_xxx attribute is known so there is no need
for per-table nest or even the OF_DPA_INFO nest.  OF_DPA_INFO nest isn't adding
any value to the command structure.

Signed-off-by: Scott Feldman <sfeldma@cumulusnetworks.com>
Acked-by: Jiri Pirko <jiri@resnulli.us>