rocker: allow for short keys on flow table match
In the flow table match algo, there was an optimization to skip over flows when
the key length of the flow doesn't match the packet match key length.  The
check should skip only when flow key length is greater than packet match
length, to allow for flow keys with partial info.  For example, consider these
two bridge table flows:

	1) vlan 100 -> DLF VLAN bridging
	2) vlan 100, dst_mac 00:01:02:03:04:05 -> Unicast VLAN Briding

The key length for 1) is shorter than 2), but both should be matched against
the packet since the packet includes both vlan and dst_mac.

Signed-off-by: Scott Feldman <sfeldma@cumulusnetworks.com>
Acked-by: Jiri Pirko <jiri@resnulli.us>