rocker: don't print err msg when stripping vlan
When flooding a pkt to multiple ports, we'll call vlan strip to remove vlan
when sending to first port, and continue to call vlan strip on pkt for
subsequent ports.  Don't print an err msg when attempting to strip a missing
vlan, which would previously happen for all sends after the first port.

Signed-off-by: Scott Feldman <sfeldma@cumulusnetworks.com>
Acked-by: Jiri Pirko <jiri@resnulli.us>