rocker: add flow tbl op for hit
We had an op for "miss", but not for "hit" on a flow table match.  (Of, there
was a "hit_no_goto" op used in case the goto tbl action was missing from a flow
match, but that a special case).  We need a "hit" op for src-mac learning
in subsequent patches.

Signed-off-by: Scott Feldman <sfeldma@gmail.com>

Acked-by: Jiri Pirko <jiri@resnulli.us>