rocker: allow flood group to reference non-existing l2 interface groups
This allows the driver to populate l2 interface groups as ports become enabled
(STP OK) and delete them when port is disabled.  The flood group will for the
bridged VLAN will reference l2 interfaces for all the ports in the VLAN, but
only active l2 interface groups will egress pkts.

Signed-off-by: Scott Feldman <sfeldma@gmail.com>