rocker: collapse rocker_flow into rocker_of_dpa
Originally , rocker_flow.[ch] was split out from of-dpa code to abstract
away non-of-dpa parts of the flow processing.  But too many of-dpa'isms
crept into flow.[ch] making it not usable for non-of-dpa flow processing.

So simplify the code base and remove flow.[ch].  If another flow processing
engine is introduced to rocker, then we can revisit and bust out the shared
code into flow.[ch].

Signed-off-by: Scott Feldman <sfeldma@gmail.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>